Begin implementing platform, p. backends, p. independence
Cursor, span, selection are now using platform-independent rectangle
implementations.

Platform class gives access to a stream of messages (locally) or a
server/socket (not implemented yet). Messages describe what the backend
needs to do, e.g. acquire rect, draw rect, etc.

This is a huge, humongous, humomongous performance regression, mainly
due to the fact that Draw messages bombard the backend rather than
notifying it on change.

Also, an `acquire`/`release` system is required for all components
that depend on Platform rects et al., with manual release and acquire
of rectangles.

On RectAcquire, the backend creates (or reuses a previously released)
rect. This is implementation-defined: if the backend has no rect class,
it can ignore RectAcquire messages. ID is attached to the RectAcquire
message, by which the rect will be identified in forthcoming messages.

On RectDraw, the backend should draw the rectangle. ID of the rectangle
is given in case RectAcquire is supported by the backend, plus rectangle
parameters like fill color, x, y, width, height.

On RectRelease, ID is released (deallocated, dropped, destroyed), and rect
object (if any) is moved into released rects, so that anyone wanting a rect
can get it rather than allocating an entirely new rect object.